Fixed issues with connect and createConnection
1) The Mongoose.connect returns self so there is no need to create
separate connection variable that gets passed to the handleConnection
function in the callback. This resolved the potential timing issue with
the callback referring back to the surrounding functions return value..

2) The Mongoose.createConnection returns the actual connection and
referring back to the connection within the callback can cause timing
issues if the callback is called on the same tick (which it is in some
error cases etc). Wrapping the handleConnection in process.nextTick
should fix the issue.